# yukka-ui/src/openApi/openApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Spring Security OpenApi",
    "description": "Dokumentacja OpenApi dla Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Łukasz Domin",
      "email": "lukaszdomin@interia.pl"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "0.3"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/rosliny/{nazwa-lacinska}": {
      "put": {
        "tags": [
          "Roslina"
        ],
        "operationId": "updateRoslina",
        "parameters": [
          {
            "name": "nazwa-lacinska",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoslinaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoslinaResponse"
                }
              }
            }
          }
        }
      }
    },
    "/rosliny/{nazwa-lacinska}/obraz": {
      "put": {
        "tags": [
          "Roslina"
        ],
        "operationId": "updateRoslinaObraz",
        "parameters": [
          {
            "name": "nazwa-lacinska",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/rosliny": {
      "post": {
        "tags": [
          "Roslina"
        ],
        "operationId": "saveRoslina",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "request"
                ],
                "type": "object",
                "properties": {
                  "request": {
                    "$ref": "#/components/schemas/RoslinaRequest"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoslinaResponse"
                }
              }
            }
          }
        }
      }
    },
    "/rosliny/wlasciwosciQuery": {
      "post": {
        "tags": [
          "Roslina"
        ],
        "operationId": "getWlasciwosciCountFromQuery",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoslinaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "uniqueItems": true,
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WlasciwoscKatalogResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/rosliny/szukaj": {
      "post": {
        "tags": [
          "Roslina"
        ],
        "operationId": "findAllRoslinyWithParameters",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 12
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoslinaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseRoslinaResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/rosliny/roslina-id/{roslina-id}": {
      "get": {
        "tags": [
          "Roslina"
        ],
        "operationId": "findByRoslinaId",
        "parameters": [
          {
            "name": "roslina-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoslinaResponse"
                }
              }
            }
          }
        }
      }
    },
    "/rosliny/nazwa-lacinska/{nazwa-lacinska}": {
      "get": {
        "tags": [
          "Roslina"
        ],
        "operationId": "findByNazwaLacinska",
        "parameters": [
          {
            "name": "nazwa-lacinska",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoslinaResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/rosliny/{roslina-id}": {
      "delete": {
        "tags": [
          "Roslina"
        ],
        "operationId": "deleteRoslina",
        "parameters": [
          {
            "name": "roslina-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RoslinaRequest": {
        "required": [
          "nazwa",
          "nazwaLacinska",
          "opis",
          "wlasciwosci",
          "wysokoscMax",
          "wysokoscMin"
        ],
        "type": "object",
        "properties": {
          "roslinaId": {
            "type": "string"
          },
          "nazwa": {
            "type": "string"
          },
          "opis": {
            "type": "string"
          },
          "obraz": {
            "type": "string"
          },
          "wysokoscMin": {
            "maximum": 200,
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "wysokoscMax": {
            "maximum": 200,
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "wlasciwosci": {
            "maxItems": 100,
            "minItems": 0,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WlasciwoscWithRelations"
            }
          },
          "nazwaLacinska": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          }
        }
      },
      "WlasciwoscWithRelations": {
        "type": "object",
        "properties": {
          "etykieta": {
            "type": "string"
          },
          "nazwa": {
            "type": "string"
          },
          "relacja": {
            "type": "string"
          }
        }
      },
      "RoslinaResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "roslinaId": {
            "type": "string"
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "nazwa": {
            "type": "string"
          },
          "nazwaLacinska": {
            "type": "string"
          },
          "opis": {
            "type": "string"
          },
          "wysokoscMin": {
            "type": "number",
            "format": "double"
          },
          "wysokoscMax": {
            "type": "number",
            "format": "double"
          },
          "obraz": {
            "type": "string",
            "format": "byte"
          },
          "autor": {
            "type": "string"
          },
          "roslinaUzytkownika": {
            "type": "boolean"
          },
          "grupy": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "formy": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "gleby": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "koloryLisci": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "koloryKwiatow": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "kwiaty": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "odczyny": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "okresyKwitnienia": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "okresyOwocowania": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "owoce": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "podgrupa": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pokroje": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "silyWzrostu": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "stanowiska": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "walory": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "wilgotnosci": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "zastosowania": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "zimozielonosci": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "NazwaLiczbaRoslin": {
        "type": "object",
        "properties": {
          "nazwa": {
            "type": "string"
          },
          "liczbaRoslin": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "WlasciwoscKatalogResponse": {
        "type": "object",
        "properties": {
          "etykieta": {
            "type": "string"
          },
          "nazwyLiczbaRoslin": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NazwaLiczbaRoslin"
            }
          }
        }
      },
      "PageResponseRoslinaResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoslinaResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Authorization Bearer JWT",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}